refactor: reorg internal plugins
