+++ packages/core/src/internal-plugin/commands.ts
@@ -1,5 +1,5 @@
 /* Copyright 2021, Milkdown by Mirone. */
-import { createContainer, createSlice, createTimer, Ctx, MilkdownPlugin, Slice, Timer } from '@milkdown/ctx';
+import { createContainer, createSlice, createTimer, MilkdownPlugin, Slice, Timer } from '@milkdown/ctx';
 import { callCommandBeforeEditorView } from '@milkdown/exception';
 import type { Command } from '@milkdown/prose';
