+++ packages/core/src/internal-plugin/index.ts
@@ -4,14 +4,7 @@ export * from './config';
 export * from './editor-state';
 export * from './editor-view';
 export * from './init';
-export * from './input-rules';
-export * from './keymap';
-export * from './mark-factory';
-export * from './node-factory';
-export * from './node-view';
 export * from './parser';
-export * from './prose-plugin-factory';
-export * from './remark-plugin-factory';
 export * from './schema';
 export * from './serializer';
 export * from './theme';
