--- packages/core/src/internal-plugin/prose-plugin-factory.ts
@@ -1,5 +1,0 @@
-/* Copyright 2021, Milkdown by Mirone. */
-import { createSlice } from '@milkdown/ctx';
-import type { Plugin } from '@milkdown/prose';
-
-export const prosePluginsCtx = createSlice<Plugin[]>([], 'prosePlugins');
